Refactor : article list component 분리
