+++ src/Router.js
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import { PageLayout } from '@common/PageLayout';
-import BoardList from './components/pages/articles/ArticleList/BoardList';
+import BoardList from './components/pages/articles/ArticleListPage';
 import ErrorPage from './components/pages/Error';
 import Login from './components/pages/login/Login';
 import Main from './components/pages/main/Main';
@@ -9,7 +9,7 @@ import Mypage from './components/pages/Mypage';
 import Auth from './components/pages/Auth';
 import Donation from './components/pages/Donation';
 import Writing from './components/pages/writing';
-import ArticleDetail from './components/pages/articles/ArticleDetail';
+import ArticleDetail from './components/pages/articles/ArticleDetailPage';
 
 const Router = () => {
   return (
